SONAR-6377 available plugins WS should sort plugins explictly
